requirements.txt
Flask==2.3.3
gTTS==2.3.2
SpeechRecognition==3.11.0
RPi.GPIO==0.7.1
# pyaudio - Install via system package: sudo apt-get install python3-pyaudio
